refactor: break circular dependency between flow-card-setup and flow-category-renderer
Extract cleanupAllDragState() and its helper clearIndicators() into a new
module src/utils/flow-drag-cleanup.js so that flow-card-setup.js no longer
imports from flow-category-renderer.js, breaking the circular dependency.

Closes #337
